uint64_t is not available under RHEL without messing with includes, so we use unsigned long long int.

+++ ptolemy/actor/lib/fmi/maHybrid/includes/fmi2TypesPlatform.h
@@ -106,7 +106,8 @@
    typedef char            fmi2Char;
    typedef const fmi2Char* fmi2String;
    typedef char            fmi2Byte;
-   typedef uint64_t        fmi2IntegerTime;
+   // uint64_t is not available under RHEL without messing with includes, so we use unsigned long long int.
+   typedef unsigned long long int fmi2IntegerTime;
 
 /* Values for fmi2Boolean  */
 #define fmi2True  1
